권한관리 상세페이지 - 탭별로 메뉴 권한 수정 및 Upsert, 리스트 페이지 - 연결된 관리자가 있다면 삭제 불가처리

+++ src/main/java/com/example/junguniv_bb/domain/authlevel/service/AuthLevelService.java
@@ -159,7 +159,14 @@ public class AuthLevelService {
         AuthLevel authLevelPS = authLevelRepository.findById(id)
                 .orElseThrow(() -> new Exception400(ExceptionMessage.NOT_FOUND_AUTH_LEVEL.getMessage()));
 
+        
+        // 연결된 관리자 수 확인
+        if(authLevelPS.getAuthLevel() > 0) {
+            throw new Exception400(ExceptionMessage.EXISTS_LINKED_MEMBER_AUTH_LEVEL.getMessage());
+        }
+
         Long authLevelValue = authLevelPS.getAuthLevel();
+
 
         // 관련 ManagerAuth 삭제
         managerAuthRepository.deleteAllByAuthLevel(authLevelValue);
